modify entry when saving for nth time
Previously I used replace functionality of isar's unique indexes. That
meant that saving an entry with a key already present in the database would
remove the old version of the given entry and save a new one. That meant
a new id was created for every change made to an entry.